src/requirements.txt
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
distlib==0.4.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.124.2
filelock==3.20.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.15
idna==3.11
iniconfig==2.3.0
kombu==5.6.1
Mako==1.3.10
MarkupSafe==3.0.3
nodeenv==1.9.1
packaging==25.0
phonenumbers==9.0.20
Pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.0
prompt_toolkit==3.0.52
pwdlib==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.8.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-usernames==1.1.0
PyYAML==6.0.3
redis==7.1.0
ruff==0.14.8
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.38.0
vine==5.1.0
virtualenv==20.35.4
wcwidth==0.2.14
